renamings and remove pageSize from ss reader
