+++ dbsnp-importer/src/main/java/uk/ac/ebi/eva/dbsnpimporter/configuration/ImportVariantsJobConfiguration.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package uk.ac.ebi.eva.dbsnpimporter.configurations;
+package uk.ac.ebi.eva.dbsnpimporter.configuration;
 
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -31,7 +31,7 @@ import org.springframework.context.annotation.Configuration;
 import org.springframework.context.annotation.Import;
 import org.springframework.context.annotation.Scope;
 
-import static uk.ac.ebi.eva.dbsnpimporter.configurations.ImportVariantsStepConfiguration.IMPORT_VARIANTS_STEP_BEAN;
+import static uk.ac.ebi.eva.dbsnpimporter.configuration.ImportVariantsStepConfiguration.IMPORT_VARIANTS_STEP_BEAN;
 
 @Configuration
 @EnableBatchProcessing
